requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
nextcord==2.4.2
Pillow==9.5.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.31.0
typing_extensions==4.6.2
urllib3==2.0.2
yarl==1.9.2
